# backend/openapi.yml
openapi: 3.0.3
info:
  title: CapeConnect API
  version: 1.1.0
  description: >
    Live backend contract for the current CapeConnect static frontend and
    supporting admin tools. Authentication uses opaque bearer session tokens,
    not JWTs.
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        operator:
          type: string
        product_type:
          type: string
        product_name:
          type: string
        journeys_included:
          type: integer
          nullable: true
        journeys_used:
          type: integer
        route_from:
          type: string
          nullable: true
        route_to:
          type: string
          nullable: true
        amount_cents:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [PAID, USED, EXPIRED, REFUNDED]
        purchased_at:
          type: string
          format: date-time
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
          nullable: true
        payment_method:
          type: string
        card_alias:
          type: string
          nullable: true
        meta:
          type: object
        qrCode:
          type: string
          nullable: true
        qrCodeSVG:
          type: string
          nullable: true
    Route:
      type: object
      properties:
        id:
          type: string
        operator:
          type: string
        route_code:
          type: string
          nullable: true
        route_name:
          type: string
        from_stop_name:
          type: string
          nullable: true
        to_stop_name:
          type: string
          nullable: true
        active:
          type: boolean
...
paths:
...
  /api/tickets:
    get:
      tags: [Tickets]
      summary: List authenticated user tickets
      security:
        - bearerAuth: []
      parameters:
        - name: operator
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Ticket history
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
    post:
      tags: [Tickets]
      summary: Create a paid ticket after fare validation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operator, productType, productName, amountCents, paymentMethod]
              properties:
                operator:
                  type: string
                productType:
                  type: string
                productName:
                  type: string
                journeysIncluded:
                  type: integer
                routeFrom:
                  type: string
                routeTo:
                  type: string
                amountCents:
                  type: integer
                currency:
                  type: string
                validFrom:
                  type: string
                  format: date-time
                validUntil:
                  type: string
                  format: date-time
                paymentMethod:
                  type: string
                cardAlias:
                  type: string
                meta:
                  type: object
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    $ref: '#/components/schemas/Ticket'
  /api/tickets/{id}/use:
    post:
      tags: [Tickets]
      summary: Mark a ticket as used or consume one journey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated ticket
  /api/tickets/verify-qr:
    post:
      tags: [Tickets]
      summary: Verify a ticket QR payload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrData]
              properties:
                qrData:
                  type: string
      responses:
        '200':
          description: Verification result
...
  /api/routes:
    get:
      tags: [Routes]
      summary: List active routes
      parameters:
        - name: operator
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Route list
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Route'
...